# Improve `ContainerTypeValidate` performance for object types